OV-421: * load user templates in creations section
