+++ frontend/src/bundles/template/components/templates-section/templates-section.tsx
@@ -8,41 +8,25 @@ import {
     Input,
     InputGroup,
     InputLeftElement,
-    Loader,
-    Overlay,
     Select,
     SimpleGrid,
 } from '~/bundles/common/components/components.js';
-import { EMPTY_VALUE } from '~/bundles/common/constants/constants.js';
-import { DataStatus } from '~/bundles/common/enums/data-status.enum.js';
 import {
-    useAppDispatch,
     useAppForm,
-    useAppSelector,
     useCallback,
-    useEffect,
     useState,
 } from '~/bundles/common/hooks/hooks.js';
 import { IconName } from '~/bundles/common/icons/icons.js';
-import { actions as studioActions } from '~/bundles/studio/store/studio.js';
+import { type Template } from '~/bundles/template/types/types.js';
 
 import { TemplateCard } from '../template-card/template-card.js';
 import { DEFAULT_TEMPLATE_PAYLOAD } from './constants.js';
 
-const TemplatesSection: React.FC = () => {
-    const dispatch = useAppDispatch();
+type Properties = {
+    templates: Template[];
+};
 
-    const { templates, dataStatus } = useAppSelector(({ studio }) => studio);
-
-    useEffect(() => {
-        if (templates.public.length === EMPTY_VALUE) {
-            void dispatch(studioActions.loadPublicTemplates());
-        }
-        if (!templates.isUserLoaded) {
-            void dispatch(studioActions.loadUserTemplates());
-        }
-    }, [dispatch, templates]);
-
+const TemplatesSection: React.FC<Properties> = ({ templates }) => {
     const [selectedFormat, setSelectedFormat] = useState<
         'landscape' | 'portrait' | null
     >(null);
@@ -61,10 +45,6 @@ const TemplatesSection: React.FC = () => {
 
     return (
         <Box padding="17px 0">
-            <Overlay isOpen={dataStatus === DataStatus.PENDING}>
-                <Loader />
-            </Overlay>
-
             <Flex
                 alignItems="center"
                 marginBottom="9px"
@@ -134,10 +114,7 @@ const TemplatesSection: React.FC = () => {
                 </Flex>
             </Flex>
             <SimpleGrid columns={{ sm: 2, md: 3, lg: 4 }} spacing="20px">
-                {templates.public.map(({ id, ...template }) => (
-                    <TemplateCard key={id} {...template} id={id} />
-                ))}
-                {templates.user.map(({ id, ...template }) => (
+                {templates.map(({ id, ...template }) => (
                     <TemplateCard key={id} {...template} id={id} />
                 ))}
             </SimpleGrid>
