+++ frontend/src/bundles/template/pages/templates.tsx
@@ -4,13 +4,24 @@ import {
     Flex,
     Header,
     Icon,
+    Loader,
+    Overlay,
     Sidebar,
     Text,
 } from '~/bundles/common/components/components.js';
 import { useCollapse } from '~/bundles/common/components/sidebar/hooks/use-collapse.hook.js';
+import { EMPTY_VALUE } from '~/bundles/common/constants/constants.js';
+import { DataStatus } from '~/bundles/common/enums/data-status.enum.js';
 import { AppRoute } from '~/bundles/common/enums/enums.js';
-import { useCallback, useNavigate } from '~/bundles/common/hooks/hooks.js';
+import {
+    useAppDispatch,
+    useAppSelector,
+    useCallback,
+    useEffect,
+    useNavigate,
+} from '~/bundles/common/hooks/hooks.js';
 import { IconName } from '~/bundles/common/icons/icons.js';
+import { actions as studioActions } from '~/bundles/studio/store/studio.js';
 import {
     CreationsSection,
     TemplatesSection,
@@ -19,9 +30,21 @@ import {
 import styles from './styles.module.css';
 
 const Templates: React.FC = () => {
-    const { isCollapsed } = useCollapse();
-
     const navigate = useNavigate();
+    const dispatch = useAppDispatch();
+
+    const { templates, dataStatus } = useAppSelector(({ studio }) => studio);
+
+    useEffect(() => {
+        if (templates.public.length === EMPTY_VALUE) {
+            void dispatch(studioActions.loadPublicTemplates());
+        }
+        if (!templates.isUserLoaded) {
+            void dispatch(studioActions.loadUserTemplates());
+        }
+    }, [dispatch, templates]);
+
+    const { isCollapsed } = useCollapse();
 
     const handleClick = useCallback(() => {
         navigate(AppRoute.STUDIO);
@@ -29,6 +52,10 @@ const Templates: React.FC = () => {
 
     return (
         <Box bg="background.900" minHeight="100vh">
+            <Overlay isOpen={dataStatus === DataStatus.PENDING}>
+                <Loader />
+            </Overlay>
+
             <Header />
             <Sidebar>
                 <Box
@@ -56,8 +83,11 @@ const Templates: React.FC = () => {
                             ></Button>
                         </Flex>
 
-                        <CreationsSection onClick={handleClick} />
-                        <TemplatesSection />
+                        <CreationsSection
+                            templates={templates.user}
+                            onClick={handleClick}
+                        />
+                        <TemplatesSection templates={templates.public} />
                     </Flex>
                 </Box>
             </Sidebar>
