+++ frontend/src/bundles/template/types/template.type.ts
@@ -1,8 +1,1 @@
-type Template = {
-    id: string;
-    name: string;
-    url: string | null;
-    previewUrl: string;
-};
-
-export { type Template };
+export { type Template } from 'shared';
